+++ dart/constraint/ContactConstraint.cpp
@@ -114,14 +114,24 @@ ContactConstraint::ContactConstraint(
   //----------------------------------------------
   // TODO(JS): Assume the frictional coefficient can be changed during
   //           simulation steps.
-  // Update mFrictionalCoff
-  const double frictionCoeffA = computeFrictionCoefficient(shapeNodeA);
-  const double frictionCoeffB = computeFrictionCoefficient(shapeNodeB);
+  // Update mFrictionCoeff
+  const double primaryFrictionCoeffA =
+                       computePrimaryFrictionCoefficient(shapeNodeA);
+  const double primaryFrictionCoeffB =
+                       computePrimaryFrictionCoefficient(shapeNodeB);
+  const double secondaryFrictionCoeffA =
+                       computeSecondaryFrictionCoefficient(shapeNodeA);
+  const double secondaryFrictionCoeffB =
+                       computeSecondaryFrictionCoefficient(shapeNodeB);
 
   // TODO(JS): Consider providing various ways of the combined friction or
   // allowing to override this method by a custom method
-  mFrictionCoeff = std::min(frictionCoeffA, frictionCoeffB);
-  if (mFrictionCoeff > DART_FRICTION_COEFF_THRESHOLD)
+  mPrimaryFrictionCoeff =
+      std::min(primaryFrictionCoeffA, primaryFrictionCoeffB);
+  mSecondaryFrictionCoeff =
+      std::min(secondaryFrictionCoeffA, secondaryFrictionCoeffB);
+  if (mPrimaryFrictionCoeff > DART_FRICTION_COEFF_THRESHOLD ||
+      mSecondaryFrictionCoeff > DART_FRICTION_COEFF_THRESHOLD)
   {
     mIsFrictionOn = true;
 
@@ -387,13 +397,13 @@ void ContactConstraint::getInformation(ConstraintInfo* info)
     assert(info->findex[0] == -1);
 
     // Upper and lower bounds of tangential direction-1 impulsive force
-    info->lo[1] = -mFrictionCoeff;
-    info->hi[1] = mFrictionCoeff;
+    info->lo[1] = -mPrimaryFrictionCoeff;
+    info->hi[1] = mPrimaryFrictionCoeff;
     info->findex[1] = 0;
 
     // Upper and lower bounds of tangential direction-2 impulsive force
-    info->lo[2] = -mFrictionCoeff;
-    info->hi[2] = mFrictionCoeff;
+    info->lo[2] = -mSecondaryFrictionCoeff;
+    info->hi[2] = mSecondaryFrictionCoeff;
     info->findex[2] = 0;
 
     //------------------------------------------------------------------------
@@ -700,6 +710,48 @@ double ContactConstraint::computeFrictionCoefficient(
   }
 
   return dynamicAspect->getFrictionCoeff();
+}
+
+//==============================================================================
+double ContactConstraint::computePrimaryFrictionCoefficient(
+    const dynamics::ShapeNode* shapeNode)
+{
+  assert(shapeNode);
+
+  auto dynamicAspect = shapeNode->getDynamicsAspect();
+
+  if (dynamicAspect == nullptr)
+  {
+    dtwarn << "[ContactConstraint] Attempt to extract "
+           << "primary friction coefficient "
+           << "from a ShapeNode that doesn't have DynamicAspect. The default "
+           << "value (" << DART_DEFAULT_FRICTION_COEFF << ") will be used "
+           << "instead.\n";
+    return DART_DEFAULT_FRICTION_COEFF;
+  }
+
+  return dynamicAspect->getPrimaryFrictionCoeff();
+}
+
+//==============================================================================
+double ContactConstraint::computeSecondaryFrictionCoefficient(
+    const dynamics::ShapeNode* shapeNode)
+{
+  assert(shapeNode);
+
+  auto dynamicAspect = shapeNode->getDynamicsAspect();
+
+  if (dynamicAspect == nullptr)
+  {
+    dtwarn << "[ContactConstraint] Attempt to extract "
+           << "secondary friction coefficient "
+           << "from a ShapeNode that doesn't have DynamicAspect. The default "
+           << "value (" << DART_DEFAULT_FRICTION_COEFF << ") will be used "
+           << "instead.\n";
+    return DART_DEFAULT_FRICTION_COEFF;
+  }
+
+  return dynamicAspect->getSecondaryFrictionCoeff();
 }
 
 //==============================================================================